+++ frontend/src/app/components/search/search-bar/search-bar.component.js
@@ -1,0 +1,12 @@
+var searchBar = {
+  bindings: {
+    onAdd: '&',
+    onSearch: '&'
+  },
+  templateUrl: './search-bar.html',
+  controller: 'SearchBarController'
+};
+
+angular
+  .module('components.search')
+  .component('searchBar', searchBar);