+++ frontend/src/app/components/search/search-pantry/search-pantry.controller.js
@@ -1,0 +1,7 @@
+function SearchPantryController(PantryModel){
+  var ctrl = this;
+}
+
+angular
+  .module('components.search')
+  .controller('SearchPantryController', SearchPantryController);