+++ frontend/src/app/components/search/search/search.component.js
@@ -1,0 +1,18 @@
+var search = {
+  templateUrl: './search.html',
+  controller: 'SearchController'
+};
+
+angular
+  .module('components.search')
+  .component('search', search)
+  .config(function ($stateProvider, $urlRouterProvider){
+    $stateProvider
+      .state('search', {
+        parent: 'app',
+        redirectTo: 'results',
+        url: '/search',
+        component: 'search'
+      })
+    $urlRouterProvider.otherwise('/app/search')
+});